Adding datetimepicker
Changed the annotation for `Consumer.init` so the pattern matching could
work without any hacky fiddling.

For example, when returning a Task from `init`, as most pages do, I can't
set the `page` field of the model. B/c of this, the ConsumerMsg will
match, but the page will be seen as `Blank` b/c I was not able to set
the page in the model when Consumer.init returned.