Adding Hold functionality
This is the solution to the problem of when an IC adds a time entry on
the same date and service code as a previous entry.

The first person, with knowledge that another IC will be adding the same
that same day, will need to check the 'Hold' box and this will alert the
system that there is to be another expected entry for that same consumer
for the same day and service.

Note that the `hold` db field is only so Leta can query by it, I don't
need to use that to search for previous holds (I search on `consumer`,
`serviceDate` and `serviceCode`).