Added realSpecialist field
This will enable the server to be able to distinguish when an admin
is sending a POST to create a bill sheet, which probably will have
a different specialist in the `specialist` field.

Since admins can back date a bill sheet, the server must know when
an admin is trying to do so to allow the request.

This is analogous to linux' real user id.

See `isLegalDate` method.